requirements.txt
slack-sdk==3.23.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1